Library Catalog –
Create a book class with a title, page count, ISBN and whether or not it
is checked out or not. Manage a collection of various books and allow
the user to check out books or return books. For added complexity
generate a report of those books overdue and any fees. Also allow users
to put books on reserve.